+++ src/main/java/uk/gov/hmcts/sptribs/config/OpenAPIConfiguration.java
@@ -13,7 +13,7 @@ public class OpenAPIConfiguration {
     @Bean
     public OpenAPI openAPI() {
         return new OpenAPI()
-            .info(new Info().title("sptribs-dss-backend")
+            .info(new Info().title("sptribs-case-api")
                       .description("Create & update case and Upload and delete docs ")
                       .version("v0.0.1")
                       .license(new License().name("MIT").url("https://opensource.org/licenses/MIT")))
