--- src/main/java/uk/gov/hmcts/sptribs/edgecase/model/LanguagePreference.java
@@ -1,16 +1,0 @@
-package uk.gov.hmcts.sptribs.edgecase.model;
-
-public enum LanguagePreference {
-    ENGLISH("english"),
-    WELSH("welsh");
-
-    private final String code;
-
-    LanguagePreference(String code) {
-        this.code = code;
-    }
-
-    public String getCode() {
-        return code;
-    }
-}
